Separate contract utilities
